+++ CHANGELOG.md
@@ -1,24 +1,41 @@
 # @agentuity/sdk Changelog
 
-## 0.0.154
+## [0.0.156] - 2025-10-15
+
+### Added
+
+- Support for listing streams with flexible filtering and pagination options ([#200](https://github.com/agentuity/sdk-js/pull/200))
+- Support for deleting streams by ID ([#200](https://github.com/agentuity/sdk-js/pull/200))
+
+### Changed
+
+- Updated telemetry logging to be more generic ([#199](https://github.com/agentuity/sdk-js/pull/199))
+
+## [0.0.155] - 2025-10-14
+
+### Patch Changes
+
+- Added functionality to send email.
+
+## [0.0.154] - 2025-10-08
 
 ### Patch Changes
 
 - Added prompt tracking data to aisdk
 
-## 0.0.153
+## [0.0.153] - 2025-10-08
 
 ### Patch Changes
 
 - enable custom user otel logger
 
-## 0.0.152
+## [0.0.152] - 2025-10-06
 
 ### Patch Changes
 
 - Fixed issue with otel headers not propagating for outbound services
 
-## 0.0.151
+## [0.0.151] - 2025-10-06
 
 ### Patch Changes
 
@@ -809,6 +826,12 @@ All notable changes to this project will be documented in this file.
 
 - Refactor the JS SDK to better support new workflow
 
+[0.0.156]: https://github.com/agentuity/sdk-js/compare/v0.0.155...v0.0.156
+[0.0.155]: https://github.com/agentuity/sdk-js/compare/v0.0.154...v0.0.155
+[0.0.154]: https://github.com/agentuity/sdk-js/compare/v0.0.153...v0.0.154
+[0.0.153]: https://github.com/agentuity/sdk-js/compare/v0.0.152...v0.0.153
+[0.0.152]: https://github.com/agentuity/sdk-js/compare/v0.0.151...v0.0.152
+[0.0.151]: https://github.com/agentuity/sdk-js/compare/v0.0.150...v0.0.151
 [0.0.150]: https://github.com/agentuity/sdk-js/compare/v0.0.149...v0.0.150
 [0.0.149]: https://github.com/agentuity/sdk-js/compare/v0.0.148...v0.0.149
 [0.0.148]: https://github.com/agentuity/sdk-js/compare/v0.0.147...v0.0.148